Move extracted functions to under library
